Many menu improvements:
Menu could be created without items
Menu could be enabled/disabled (no events will be accepted when it is disabled)
It is possible to change its priority on runtime
kCCMenuTouchPriority -> kCCMenuHandlerPriority


Former-commit-id: 00a56482acf473520547805ef0a4f22b18832803

+++ tests/ParticleTest.m
@@ -1792,7 +1792,7 @@ Class restartAction()
 	{
 		
 		CCLOG(@"remove random system");
-		uint rand = arc4random() % ([[batchNode_ children] count] - 1);
+		NSUInteger rand = arc4random() % ([[batchNode_ children] count] - 1);
 		[batchNode_ removeChild:[[batchNode_ children] objectAtIndex:rand] cleanup:YES];
 		
 		CCParticleSystemQuad *particleSystem = [CCParticleSystemQuad particleWithFile:@"Particles/Spiral.plist"];
